flock_ticket_ingester/pyproject.toml
[tool.poetry]
name = "flock-ticket-ingester"
version = "0.1.0"
description = ""
authors = ["Adam Yodinsky <27074934+adamyodinsky@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "flock_ticket_ingester"}]

[tool.poetry.dependencies]
python = "^3.10"
pika = "^1.3.2"
fastapi = {extras = ["all"], version = "^0.95.2"}
py-trello = "^0.19.0"
pydantic = "^1.10.8"
click = "^8.1.3"
flock-common = {path = "../flock_common", develop = true}
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
